+++ tests/resources/test_generic_resoure.py
@@ -82,11 +82,7 @@ def test_generic_resource_with_bad_json_as_string_is_converted_to_a_string_prope
 
 
 def test_parse_generic_resource_without_properties():
-    resource = GenericResource.parse_obj(
-        {
-            "Type": "AWS::SNS::Topic",
-        }
-    )
+    resource = GenericResource.parse_obj({"Type": "AWS::SNS::Topic"})
     assert isinstance(resource, GenericResource)
     assert resource.Properties is None
     assert resource.Type == "AWS::SNS::Topic"